reading las functionalities included in las_handler
